calculate ttl

+++ src/main/java/eu/dissco/backend/domain/MasJobRecordFull.java
@@ -17,7 +17,7 @@ public record MasJobRecordFull(
     Instant timeCompleted,
     JsonNode annotations,
     boolean batchingRequested,
-    Integer timeToLive,
+    Long timeToLive,
     ErrorCode errorCode) {
 
 }
