+++ src/main/java/eu/dissco/backend/domain/MasJobRequest.java
@@ -3,5 +3,5 @@ package eu.dissco.backend.domain;
 public record MasJobRequest(
     String masId,
     boolean batching,
-    Integer timeToLive) {
+    Long timeToLive) {
 }
